requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.1.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bertopic==0.16.2
blinker==1.9.0
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.87
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
customtkinter==5.2.1
cycler==0.12.1
Cython==0.29.37
darkdetect==0.8.0
debugpy==1.8.1
decorator==5.1.1
Django==5.1.3
djangorestframework==3.15.2
etils==1.11.0
executing==2.0.1
filelock==3.14.0
Flask==3.1.0
flatbuffers==24.3.25
flax==0.10.2
fonttools==4.51.0
frozenlist==1.5.0
fsspec==2024.6.0
gast==0.6.0
genai==2.1.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
h5py==3.12.1
hdbscan==0.8.36
holehe==1.61
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
humanize==4.11.0
idna==3.6
imbalanced-learn==0.12.2
imblearn==0.0
importlib-metadata==6.11.0
importlib_resources==6.4.5
ipykernel==6.29.4
ipynb-py-convert==0.4.6
ipython==8.25.0
itsdangerous==2.2.0
jax==0.4.36
jaxlib==0.4.36
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.2
oauthlib==3.2.2
openai==0.27.10
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.10.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.8.1b0
Pygments==2.18.0
PyJWT==2.10.1
pynndescent==0.5.12
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.32.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
safetensors==0.4.3
Sastrawi==1.0.1
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
selenium==4.23.1
sentence-transformers==3.0.0
simplejson==3.19.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.2
stack-data==0.6.3
streamlit==1.34.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.11.0
tensorflowjs==4.22.0
tensorstore==0.1.69
termcolor==2.4.0
textblob==0.18.0.post0
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.3.3
tk==0.1.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.0
torch==2.3.0
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.41.2
trio==0.26.0
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.6
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.1.1
yarl==1.18.3
ydf==0.9.0
zipp==3.17.0
